requirements.txt
# Core ML/GNN Stack
torch>=2.0.0
torchvision>=0.15.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0

# Chemistry toolkit
rdkit>=2023.3.1

# Drug dataset access
PyTDC>=0.4.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Experiment tracking
wandb>=0.15.0

# API and demo
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
pydantic>=2.0.0

# Explainability
shap>=0.42.0
captum>=0.6.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
py3Dmol>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0
